Here’s what was wrong and what we changed.
1. MissingMethodException on SqlError
In Microsoft.Data.SqlClient 6.x, the internal SqlError constructor no longer takes (uint) for the native/win32 slot; it uses int for that parameter. The test helper was updated to pass 0 (int) instead of (uint)0, with a short comment explaining the version shift.

2. NullReferenceException on null tests
SqlTransientDetector accessed ex.Number / ex.InnerException without guarding null. That didn’t match the tests (and your null-checking rule). Both overloads now return false when the argument is null.

Result: All 15 SqlTransientDetectorTests pass.

    public static bool IsTransient(SqlException ex)
    {
        if (ex is null)
            return false;
        return ex.Number is -2 or 40613 or 40197 or 49918 or 49919 or 49920;
    }
    /// <summary>Returns <see langword="true"/> when the exception (or its inner exception) is a transient SQL or timeout error.</summary>
    public static bool IsTransient(Exception ex)
    {
        if (ex is null)
            return false;
        if (ex is SqlException sqlEx)
            return IsTransient(sqlEx);
        if (ex is TimeoutException)
            return true;
        if (ex.InnerException is SqlException innerSql)
            return IsTransient(innerSql);
        return false;
    }

+++ ArchiForge.Api.Tests/SqlTransientDetectorTests.cs
@@ -98,21 +98,24 @@ public sealed class SqlTransientDetectorTests
             args: null,
             culture: null)!;
 
+        // Microsoft.Data.SqlClient 6.x: internal ctor is
+        // (int infoNumber, byte state, byte errorClass, string server, string message, string procedure,
+        //  int lineNumber, int nativeError, Exception exception) — older builds used uint for native/win32.
         SqlError error = (SqlError)Activator.CreateInstance(
             typeof(SqlError),
             BindingFlags.NonPublic | BindingFlags.Instance,
             binder: null,
             args:
             [
-                number,        // infoNumber
-                (byte)0,       // errorState
-                (byte)0,       // errorClass
-                "server",      // server
-                "message",     // errorMessage
-                "procedure",   // procedure
-                0,             // lineNumber
-                (uint)0,       // win32ErrorCode
-                null           // exception
+                number,
+                (byte)0,
+                (byte)0,
+                "server",
+                "message",
+                "procedure",
+                0,
+                0,
+                null
             ],
             culture: null)!;
